modularized code, added workflows for linting formatting and building
