+++ .github/workflows/build.yml
@@ -1,0 +1,27 @@
+name: build
+
+on:
+  push:
+    branches:
+    - '**'
+  pull_request:
+    branches:
+    - 'main'
+
+jobs:
+  install:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - uses: pnpm/action-setup@v4
+        with:
+          version: 9
+
+      - name: Install dependencies
+        run: pnpm install
+
+      - name: build
+        run: pnpm build
