+++ .github/workflows/lint-format.yml
@@ -1,0 +1,22 @@
+name: lint, format
+
+on:
+  push:
+    branches:
+    - '**'
+  pull_request:
+    branches:
+    - 'main'
+
+jobs:
+  quality:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Setup Biome
+        uses: biomejs/setup-biome@v2
+
+      - name: Run Biome
+        run: biome ci .
