+++ jest.config.ts
@@ -1,10 +1,10 @@
 import type { Config } from "jest";
 
 const config: Config = {
-  verbose: true,
-  preset: "ts-jest",
-  testEnvironment: "node",
-  moduleFileExtensions: ["ts", "js", "node"],
+	verbose: true,
+	preset: "ts-jest",
+	testEnvironment: "node",
+	moduleFileExtensions: ["ts", "js", "node"],
 };
 
 export default config;
