+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-	"include": ["lib/**/*"],
+	"include": ["src/**/*"],
 	"compilerOptions": {
 		/* Visit https://aka.ms/tsconfig to read more about this file */
 		/* Projects */
@@ -28,7 +28,7 @@
 		"module": "commonjs" /* Specify what module code is generated. */,
 		// "rootDir": "./",                                  /* Specify the root folder within your source files. */
 		// "moduleResolution": "node10",                     /* Specify how TypeScript looks up a file from a given module specifier. */
-		"baseUrl": "./lib" /* Specify the base directory to resolve non-relative module names. */,
+		"baseUrl": "./src" /* Specify the base directory to resolve non-relative module names. */,
 		// "paths": {},                                      /* Specify a set of entries that re-map imports to additional lookup locations. */
 		// "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
 		// "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
